Consolidate alpha channel handling
