+++ src/converters/hex2rgb.ts
@@ -1,8 +1,7 @@
 import formatHex from '~/format-hex';
+import { addAlpha, extractAlphaFromHex } from '~/modules/alpha';
 import { MESSAGES } from '~/modules/constants';
-import { extractAlphaFromHex } from '~/modules/hex-utils';
 import { invariant } from '~/modules/invariant';
-import { addAlpha } from '~/modules/utils';
 import { isHex } from '~/modules/validators';
 
 import { RGB } from '~/types';
