+++ src/converters/oklch2hsl.ts
@@ -1,6 +1,7 @@
 import oklch2rgb from '~/converters/oklch2rgb';
 import rgb2hsl from '~/converters/rgb2hsl';
-import { addAlpha, extractAlpha, parseInput } from '~/modules/utils';
+import { addAlpha, extractAlpha } from '~/modules/alpha';
+import { parseInput } from '~/modules/utils';
 
 import { ConverterParameters, HSL, LCH } from '~/types';
