+++ src/converters/oklch2oklab.ts
@@ -1,6 +1,7 @@
 /* eslint-disable prefer-const */
+import { addAlpha, extractAlpha } from '~/modules/alpha';
 import { DEG2RAD } from '~/modules/constants';
-import { addAlpha, extractAlpha, parseInput, restrictValues } from '~/modules/utils';
+import { parseInput, restrictValues } from '~/modules/utils';
 
 import { ConverterParameters, LAB, LCH } from '~/types';
